+++ src/main/kotlin/caiqichang/mybatisuite/navigation/MapperFileDescription.kt
@@ -6,9 +6,7 @@ import com.intellij.util.xml.DomFileDescription
 
 class MapperFileDescription : DomFileDescription<Mapper>(Mapper::class.java, "mapper") {
 
-    val enable = true
-
     // must override, constructor and super function do not check the root tag name 
     override fun isMyFile(file: XmlFile, module: Module?) =
-        file.rootTag != null && file.rootTag?.name == "mapper" && enable
+        MapperUtil.enableXmlMapperResolving() && file.rootTag != null && file.rootTag?.name == "mapper"
 }